refactor based on codebase
